mobile/requirements.txt
flet>=0.84.0
openai>=1.0.0
loguru>=0.7.0
python-docx>=1.0.0
markdown>=3.5.0
lxml>=5.0.0
httpx>=0.28.0
tiktoken>=0.7.0
msgpack>=1.1.0
oauthlib>=3.2.0
repath>=0.9.0
